Removed the 1.0 code from the messages crate. It builds and tests pass.
